React Conversion: Refactored Frontend File Structure
