+++ .babelrc
@@ -1,3 +1,4 @@
 {
-  "presets": ["env",  "react"]
+  "presets": ["@babel/preset-env",  "@babel/preset-react"],
+  "plugins": ["@babel/plugin-transform-runtime"]
  }
